[cpp] getDrafts API
Summary:
The purpose of this API is to make it possible to fetch the drafts into the JS runtime when the app boots up so we don't have to worry about blocking reads and performance. For now we have great performance, but this may change in the future (separate SQLite thread, larger databases, etc.). And for more complicated data like `ThreadStore` / `MessageStore` I expect this approach will be necessary.

I think the next step here is to make this API asynchronous (eg. return `Promise` on the JS side), and then to actually implement `CoreDataProvider`, which will fetch the data from the C++ side and hold onto it, and keep it updated. To implement `CoreDataProvider` we will need to think a little harder about how to implement subscriptions. In the case of drafts we could just update `CoreDataProvider` and the SQLite database at the same time, but in other cases the updates won't always flow from JS, so it's probably best to start thinking about how we can update the JS store based on changes on the C++ side.

I tried to keep the "business logic" in `CommCoreModule`, but I'm not sure it's the right approach. If we filtered out empty strings in `SQLiteManager` (not just on reads, but also `DELETE FROM` on writes) we could keep the database cleaner. @karol-bisztyga, what do you think?

Test Plan: I just added some JS code to call `global.CommCoreModule.getDrafts` to make sure it works

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: KatPo, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1115