[cpp] Don't pass Runtime to DatabaseManagerInterface
Summary:
I'd like the `DatabaseManagerInterface` to be possible to use on a non-JS thread, but the `Runtime` object is supposed to only be used on the JS thread.

To accomodate this, this diff modifies our code to avoid passing the `Runtime` to the `DatabaseManagerInterface`. Luckily the `Runtime` isn't being used, so it was trivial.

If we want to allow multiple threads to use `DatabaseManagerInterface`, we could make it thread-local. The `sqlite_orm` should setup two separate `sqlite3` connections. We could do this by making the `static DatabaseManagerInterface` inside `DatabaseManager` a `folly::ThreadLocal`.

Test Plan: Make sure it compiles

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1117