[cpp] Return Promise from getDrafts
Summary:
This diff doesn't really do much... it's basically the equivalent of `setTimeout`.

Right now the time between call and `Promise` resolution is measuring as 1ms as it's all synchronous on the same thread.

I think in combination with D1074 this will make much more sense - letting the DB operation occur on a separate thread and then calling back to the JS thread with the result.

If we can make sure that we aren't causing threading problems, we could make both on-thread synchronous reads and off-thread asynchronous reads work. But we'll have to think about how we pick up changes on writes (both for app-level subscriptions, and in terms of how two connections to the same SQLite databases on two different threads work when one of the connections is writing and the other is reading).

Test Plan:
```lang=js
global.CommCoreModule.getDrafts.then(drafts => console.log(drafts));
```

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1116