[native] Create symlink to allow Android to resolve ReactCommon
Summary:
@karol-bisztyga originally set up this hack where we include a different directory on Android, and we've used it in several places in our codebase.

But it's not ideal when including files from `react-native` directly, since we would need to patch those files directly, and would break our ability to compiled `ReactAndroid` from source.

This diff solves the issue by creating a symlink with the expected directory structure that we can pass to `CMake`.

Test Plan: Make sure it compiles and runs on iOS and Android, and that drafts read / write correctly on both

Reviewers: karol-bisztyga, palys-swm

Reviewed By: karol-bisztyga, palys-swm

Subscribers: KatPo, Adrian, atul, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1134