[lib] Remove errorHandler from createRealThreadFromPendingThread
Summary: In D718 we introduced thread creation hook which had a `handleError` parameter. When it was present, all exceptions would be caught and undefined was returned instead. This caused the returned promise to be optional and that was reflected in `pendingThreadCreations` map in input state containers. In D1140 and D1139 we removed all the `handleError` usages and now we can use `Promise<string>` instead which simplifies the code.

Test Plan: Flow, tested with other diffs

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1141