[native] Don't update deviceToken in Redux unless setDeviceToken API request succeeds
Summary:
Adrian experienced an issue on Android where he wasn't getting notifs delivered for some reason. He was able to solve it by logging out and in, but it was a pretty negative experience.

I read through the code in an effort to guess at what might have happened. The fact that logging out and in again solved the issue strongly indicates that the server had the wrong device token. But the native side is supposed to keep track of the device token in Redux, and inform the server whenever it changes.

I have a guess at what happened. I noticed that currently we update the `deviceToken` on `setDeviceTokenActionTypes.started`. But what if that request fails? In that case, the client will think the server has the correct device token, but the server will still have the old one.

In this diff I update it to use `setDeviceTokenActionTypes.success`, so we only update it on the client's Redux store if it can confirm the server received the request. Until the device token is updated in Redux, on every foreground the app will try to register the new device token. So it should eventually get updated once a connection with the server can successfully be established.

Test Plan: I honestly am just relying on Flow here. I really don't have much time to test this today, and it's a pretty complicated thing to test since notifs are hard to get working in the local dev environment

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1154