[server] Make joinThread idempotent
Summary: When we will call joinThread from createThread, it might happen that user is already member of this thread. In this case we don't want to get serverError, but return 'empty' response.

Test Plan: Test if joinThread works for old and new clients, check if calling it for user that is already member of thread without idempotent set to true still throws error. Checked if there is no error when idempotent is set

Reviewers: ashoat, KatPo

Reviewed By: ashoat, KatPo

Subscribers: ashoat, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1127