[native] Don't overwrite LogInPanel with native credentials on every load
Summary:
This is a minor issue I noticed with a quick fix. Right now, every time the user opens the `LogInPanel`, the fields get set based on native credentials.

Instead, this diff makes it so the `LogInPanel` is initially set with the native credentials, but if the user changes the credentials, then goes back, and then returns to `LogInPanel`, it will keep their updates.

Test Plan: Go to `LogInPanel`, change credentials, go back, and then return to `LogInPanel` and make sure the credentials are still changed

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1161