[native] Use comm.app as server for "Internet credentials" in react-native-keychain
Summary: We use `react-native-keychain` to store the username/password combo outside of Redux. We store it with this "Internet credentials" mechanism where we specify the domain, which initially made me think it relied on some permissions. But I tested this and it seemed to work without any permissions. I think it probably can't be used outside the app.

Test Plan: Make sure that credentials still persist and delete when they should

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1184