[native] Change folder where we save media on Android from SquadCal to Comm
Summary: When the user saves an image, or when they capture a photo using `CameraModal` after having granted image-saving permissions, we save the image to a custom folder on Android. This diff changes the code to save the images into a folder named "Comm" instead of a folder named "SquadCal".

Test Plan: Ehhhhh it's probably fine

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1197