[lib] Don't propagate OPEN permissions into PERSONAL / PRIVATE children
Summary: Right now `PERSONAL` and `PRIVATE` threads can't have parents, so this diff doesn't change anything. But [in this Notion task](https://www.notion.so/commapp/Surface-to-users-that-everything-is-on-Ashoat-s-keyserver-9e442ee4bbb747bcb6f3b9203017a41e) I concluded that we'll need to move every thread to be the child of a `GENESIS` community for the soft launch. As part of that, we want to make `PERSONAL` / `PRIVATE` threads behave like `CHAT_SECRET` threads in that you can't see them unless you are in them, or you are a community admin.

Test Plan: will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1200