[native] Don't include threadPermissions.JOIN_THREAD in any role permissions blobs
Summary:
Role permissions only get conferred once you are a member, and there's no point in giving a user `JOIN_THREAD` if they're already a member.

It only makes sense for `JOIN_THREAD` to appear in a role permissions blob with a prefix, ie. for it to be used for descendants in some way.

Test Plan: Made sure `JOIN_THREAD` is only used for threads the user isn't a member of. Also will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1203