[cpp] Build patch for folly on android
Summary:
This is a patch for D1148 and should be landed after that diff.

Turned out using folly requires additional steps on android(so far we tested builds only when folly was added to the project. If any of its components' functions is called, we get errors that basically mean we miss implementations for those functions).
To be specific I had to add certain source files for compilation. After a discussion, we decided not to add an entire folly lib's contents but only those which will be used.
Also, I changed the approach for attaching `glog` and `boost` to the project.
For `folly` itself, there were two patches needed due to avoid compile errors.

Test Plan:
build android with some example code like this:
```
#include <folly/MPMCQueue.h>

folly::MPMCQueue<int> q(10);
q.write(5);
```

Reviewers: palys-swm, ashoat

Reviewed By: ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1164