[cpp] Database manager structure for the separate thread
Summary: This consists of two things: a separate thread handling and task scheduling. The thread is stored as a member of `DatabaseManager`. Enqueueing a nullptr into the queue makes it terminate smoothly. It is joined in the destructor, so we make sure we handle it properly. As of the queue: I decided to use the `UnboundedQueue` instead of the `MPMCQueue`, because I don't think we want to specify a max number of tasks that may be stored and handle a situation when it exceeds the limit. I followed the instruction that's [present in `MPMCQueue`](https://github.com/facebook/folly/blob/96b303638fc0b10bcaa029703990b2954f9ee0ff/folly/MPMCQueue.h#L119) code. I assume those are alternatives and `UnboundedQueue` is as thread safe as `MPMCQueue`. Plus, `UnboundedQueue` has far more user friendly API than `MPMCQueue`. The former has the sacred pair of `enqueue` `dequeue` and the latter contains a lot of different functions that are not really well documented and seem like an overkill for us.

Test Plan: Well, it still compiles. Upcoming diff will make use of this since here `scheduleTask` is not yet invoked anywhere.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1137