[lib] Fail all permissions checks if KNOW_OF is missing
Summary:
Before D1207, if `KNOW_OF` was missing then no permissions would be saved. But we need to store these permissions in order to be able to resolve `MEMBERSHIP` permissions for potential members, so I changed the behavior.

In order to keep behavior consistent for permission checks, this diff makes permission checks all fail if `KNOW_OF` is missing.

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1213