[server] Don't trust role permissions in database if rolesNeedUpdate
Summary:
In `updateThread` we need to fetch `defaultRolePermissions` in order to pass to `validateCandidateMembers` (see D1212).

This diff changes the behavior only in the case where `forceUpdateRoot` causes `updateRoles` to be run. In D1212 I had us lookup the permission in the database, but since we can expect `updateRoles` to update the role permissions to match `getRolePermissionBlobsForChat`, we should just check that directly. (We can't run `updateRoles` first because we're trying to do all permission checks before mutating anything.)

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1214