[cpp] Prevent android crash on debug
Summary:
This prevents android from crashing when `Debug` option is selected from the developer menu. Due to using jsi we shouldn't use that option and we should use Flipper instead, but I wanted to add a safe way back in case somebody forgets about that - so they can just turn off debugging from the same menu instead of being forced to uninstall the app.

Related notion: https://www.notion.so/commapp/Android-app-crashes-on-start-on-emulator-758ee724fecd48a6968467e0031ed528

Test Plan: Go to the developer menu on android and pick `Debug`, see that it doesn't crash but shows a warning instead.

Reviewers: palys-swm, ashoat

Reviewed By: ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1187