[native] Fix incorrect prepareFolly UP-TO-DATE status in Gradle
Summary:
I was unable to build Android for some reason, and was getting the errors that I thought @karol-bisztyga had solved with the find-and-replace `filter` operations in `prepareFolly`. After investigating it I found that the files in `android/app/build/third-party-ndk/folly` did not have the replacements applied.

I did some searching and found [this Gradle issue](https://github.com/gradle/gradle/issues/1191). By adding our find-and-replace strings as "properties" we are able to tell Gradle that when these change, it needs to re-run the tasks.

Without this diff, Gradle didn't realize it needed to re-run the tasks after getting updated due to D1164. After this task, Gradle would've known it needed to re-run the tasks, and if we ever need to change these find-and-replaces, as long as we follow this pattern Gradle should work correctly.

Test Plan: Compile Android and make sure I don't see Folly issues

Reviewers: karol-bisztyga, atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian, karol-bisztyga

Differential Revision: https://phabricator.ashoat.com/D1229