[server] Prevent changing parent thread of promoted sidebars
Summary: Currently we prevent users from changing the parent thread of a sidebar, but if they promote the sidebar into an open subthread, it becomes possible. I think it would be best to avoid allowing this for now. It has the same effect as allowing promoted sidebars to be made secret... we would have to consider a client seeing a source message but not a thread that was created from it.

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1234