[server] Delete rows for role IDs when deleting a thread
Summary: I was testing `soft-launch-migration.js` by creating fake Genesis threads, and was thereafter deleting them with `deleteThread`. I noticed that the entries in the `ids` table for the roles weren't being deleted. Tracked it down to this accidental omission.

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1252