[server] Avoid fetching updates for script viewers
Summary:
We have [code to check](https://github.com/CommE2E/comm/blob/master/server/src/creators/update-creator.js#L115) for this in `update-creator.js`, but not in `update-fetchers.js`. I was seeing a crash as a result of this (there's no way to determine a default `calendarQuery` for a script viewer because we don't know their timezone).

This diff makes sure we don't try to get updates for a script viewer, which has no way to handle updates anyways.

Test Plan: Will be tested in combination with following diffs

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1253