[native] Fix up navigation default state to reflect apps tab
Summary: This was missed in D1191 and D1210. I think we discussed fixing this in person, but I may be misremembering. If so, @atul – next time would be good to make a Notion task about it so you don't forget.

Test Plan: Make sure I can log out and back in without getting an error

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, atul, Adrian

Differential Revision: https://phabricator.ashoat.com/D1259