[lib] Include MEMBERSHIP permissions in createPendingThread
Summary:
In D1207, I updated `makePermissionsBlob` so that it would ignore `rolePermissions` if the combined `permissions` does not include `MEMBERSHIP`.

We use `makePermissionsBlob` in `createPendingThread`, and we pass in a fake `rolePermissions` there that previously didn't include `MEMBERSHIP`. As a result, the role permissions were being ignored, and since there were no `permissionsFromParent`, the result was a `null` permissions blob.

This diff adds `MEMBERSHIP` so that the permissions blob gets created correctly, and thus resolves [this Notion issue](https://www.notion.so/commapp/Permissions-problems-with-sidebars-threads-edcfa9c5906c4bd2a731e7d3cb2319b2).

Long-term the right solution is to use real permissions instead of making a fake permissions blob in `createPendingThread`. That work is tracked by [this Notion issue](https://www.notion.so/commapp/Use-correct-permissions-in-createPendingThread-a849ccf79d5d434cb4fd44e18a88ed28), but is a little complicated because clients don't currently have awareness of the whole set of permission keys.

Test Plan: Make sure pending threads show `ChatInputBar` with a working `TextInput`

Reviewers: karol-bisztyga, atul

Reviewed By: karol-bisztyga

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1265