[server] Expose last_update column of sessions table as Viewer.sessionLastUpdated
Summary: We need this in order to invalidate sessions when the server has a newer update checkpoint than the client, which generally indicates the client either failed to persist its Redux store, or that the user restored the client app from a backup. More details in [this Notion issue](https://www.notion.so/commapp/Messages-disappearing-after-send-2dcfa2b4f92647f4aa5140009d6b4682).

Test Plan: Tested in combination with the next diff

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1270