[server] Invalidate a session if the client reports an outdated update checkpoint
Summary:
In [this Notion issue](https://www.notion.so/commapp/Messages-disappearing-after-send-2dcfa2b4f92647f4aa5140009d6b4682), I discovered that if the user restores a backup, we don't correctly invalidate the session.

This diff solves that issue.

Test Plan: I set the `last_update` column in the `sessions` table on MySQL to a really high value for the active session, and then backgrounded and foregrounded the app. I confirmed that the app received a `FULL_STATE_SYNC` as a result and that the `last_update` column was corrected to a valid timestamp

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1271