[lib] Increment nextLocalID when the server delivers messages with a localID
Summary:
This diff addresses the bug described by [this Notion issue](https://www.notion.so/commapp/Messages-disappearing-after-send-2dcfa2b4f92647f4aa5140009d6b4682).

We want to make sure that the client updates its `nextLocalID` whenever it sees a message from the server with a newer `localID`.

Test Plan: I edited the `location` column of a message in the `messages` table that hadn't been delivered to a client yet to make that client think it wrote the message. Then I foregrounded the client and made sure it updated its `nextLocalID` when it saw the message

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1274