[server] Push back the launch of some features
Summary: I'm going to make 84 a patch release for SquadCal, to make sure that things don't immediately break or look weird after we run the soft launch migration. As a result we need to increment some of these version checks, since the relevant client support won't be there in 84.

Test Plan: Haven't really tested this

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1276