[server] Don't check Genesis CREATE_SUBTHREADS
Summary:
We don't want any user to be able to create open subthreads on Genesis, so we've disabled `CREATE_SUBTHREADS` (it's set as `COMMUNITY_ANNOUNCEMENT_ROOT`).

However, we hack Genesis as the parent of threads where a parent isn't specified. To keep this working, this diff makes sure we don't get `CREATE_SUBTHREADS` for Genesis.

Test Plan: Tested in combination with following diff

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1290