[lib] Fix getSingleOtherUser to ignore non-members
Summary:
This diff addresses [this Notion issue](https://www.notion.so/commapp/Invalid-friend-request-123d5fe61edf42e39ad5e6563d99caee).

This function is used for two things:

1. Extracting the other user from a `PERSONAL` thread
2. Determining whether the thread should be treated as blocked in the case that the other user is blocked (or has blocked you)

In both of these cases, with the temporary approach we're taking where all `PERSONAL` threads are children of Genesis, ashoat will be listed as a member. As a result, in order to keep the previous behavior we'll need to ignore non-members.

We'll want to revert this when we switch back to `PERSONAL` threads not having parents, so that we don't "block" threads in a community.

Test Plan: Make sure the correct buttons appears in `RelationshipPrompt` and `ThreadSettings`, and that the buttons do the right thing

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1293