[lib] Keep threads that have been just reassigned
Summary:
In `threadIsWatched` we treat every pending thread as if it is watched.
In `reassignMessagesToRealizedThreads` we move messages from pending to realized threads.
It's possible to receive an update about pending thread becoming realized before we join it. This update will make us perform the reassignment, but then in `mergeNewMessages` we will remove the new thread from the store.

Why the new thread is removed? It's because the current user is not a member of it, and is not watching it. The thread is not being watched because in `useWatchThread` we pass the pending thread and only after this update is processed, the effect would update the watched thread to realized one. To fix that we need to make sure that for this one update we keep the thread. We can do that by remembering which threads were reassigned and add them temporarily to the watched list.

Test Plan:
Modify thread-responder to create thread with one second of delay.
Login as two different users on two devices and open pending sidebar created from a message created by one of the users.
Send a message on both devices. The message from the source author should be sent earlier.
Check if pending message sent by other user is visible for the whole time.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1326