[server] Update changeRoleThreadQuery to fetch containing_thread_id
Summary:
We'll need this in order to determine if membership in the containing thread is necessary for membership in the `threadID` passed to `changeRole`.

While I was here I made the queries more strict so they would surface errors, and combine two of the cases that were basically the same (0 and -1).

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1348