[server] Update changeRole to check containing thread
Summary:
I'm getting rid of `threadPermissions.MEMBERSHIP` because it has performance issues – see [this Notion task](https://www.notion.so/commapp/Permissions-system-performance-issues-with-Genesis-as-universal-root-thread-857454cc191a41a08eb1b2981abd6f8c). I'm replacing it with the concept of a "containing thread" – parent for sidebars, community otherwise. The `containing_thread_id` column was introduced in an earlier diff.

This diff updates the logic in `changeRole` to check containing thread membership. An earlier diff removes the logic that checks `threadPermissions.MEMBERSHIP`.

In this diff I also separated the initial query into three. I think this will be more performant and it's much more readable.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1350