[native] Check if height of messages changed
Summary: Currently it's possible for a message to change its height. There is an optimization in `chat-list` which check if the number of messages is the same - if that's true, we assume that the height and scroll positions don't have to be updated. Now, there's another option: someone can create a sidebar from a message, which alters its height. I decided to compare the heights instead of simply checking `threadCreatedFromMessage` flag, so that when other options of height change are introduced, the behavior will remain correct.

Test Plan: Open a thread and send a message from other device to it. When a message is below the view, currently displayed messages shouldn't be moved. When a message is older, messages above it should be moved.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1341