[server] Update validateCandidateMembers to check containing thread membership
Summary: I'm getting rid of `threadPermissions.MEMBERSHIP` because it has performance issues – see [this Notion task](https://www.notion.so/commapp/Permissions-system-performance-issues-with-Genesis-as-universal-root-thread-857454cc191a41a08eb1b2981abd6f8c). I'm replacing it with the concept of a "containing thread" – parent for sidebars, community otherwise. The `containing_thread_id` column was introduced in an earlier diff – this diff uses that to check membership instead of `threadPermissions.MEMBERSHIP`.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1346