[server] Delete create-personal-thread.js and create-private-thread.js scripts
Summary:
Since `recalculateThreadPermissions` is going to need to fetch `containing_thread_id` off the `threads` table anyways, I figured we could have it fetch the `ThreadType` at the same time so it doesn't have to take a `threadType` parameter anymore.

However, to do that we need to make sure the `type` column in the `threads` table is updated before `recalculateThreadPermissions` is called. These two scripts don't do that, and besides that they are the only two callers to `DEPRECATED_updateRoleAndPermissions`, which I will remove in the next diff.

Test Plan: Flow

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1353