[server] Fetch ThreadType in recalculateThreadPermissions
Summary: Before this diff, `recalculateThreadPermissions` needed to get the `ThreadType` passed in. Since we need to fetch `containing_thread_id` off of the same row in the `threads` table, and since all of the callsites guarantee that the `threads` table is updated before calling `recalculateThreadPermissions`, we can simplify things by having the `ThreadType` get fetched instead of being passed in.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1352