[server] Don't skip updates in recalculateThreadPermissions if role changed
Summary: It's hypothetically possible for `permissions` not to change, but for `role` to change. For instance, we might determine that a community admin doesn't have membership in a sidebar's parent, so they must lose membership in the sidebar. This would necessitate a change in the `role` column in the `memberships` table. But if the user is an admin of the community, perhaps they have all of the same permissions for the sidebar, whether or not they are a member of its parent.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1364