[server] Reorganize error conditions in changeRole
Summary: This is a little cleaner and should result in the same behavior. The `console.warn` check is now a bit more generic, but as far as I can tell it should only be triggered when the user ask for -1 and gets 0, or vice versa.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1366