[cpp] SQLite: set single threaded mode
Summary:
As we decided to use SQLite only on the database thread, for now, we want to set the threading mode in SQLite to single-threaded.

Related notion doc: https://www.notion.so/commapp/Force-single-threading-mode-in-SQLite-77a7bd13bbf042c183923b4383ad22d5

Test Plan:
In the linked notion doc I put information on how the built files changed in the android build.
We could do some performance benchmarking but I'm not sure if that's neither necessary nor desired considering that the performance differences are probably pretty tiny.

Reviewers: palys-swm, ashoat

Reviewed By: ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1264