[server] Avoid creating unnecessary ghost membership rows
Summary:
A "ghost" membership row is one where `role = -1 AND permissions IS NULL`. These should be left over after a user leaves a thread, so that new members of the thread can have a relationship created with the old users so that the server knows to send the corresponding `UserInfo`s over.

Right now the updated thread permissions code is creating a lot of unnecessary ghost rows in cases where it shouldn't. This diff fixes up the code to avoid that by making sure we `continue;` in those cases, and also fixes a couple cases where `continue;` conditions weren't being triggered correctly.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm, atul

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1385