[server] Fix updateRoles code for creating admin role when it doesn't exist
Summary: We need to `JSON.stringify` objects before we send them to be stored in a `JSON` MySQL row. I never noticed this bug before because we previously had never converted a non-community to a community, but the new `soft-launch-migration.js` code does this.

Test Plan: Will be tested in combination with following diffs

Reviewers: palys-swm, atul

Reviewed By: palys-swm, atul

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1386