[native] Refactor database thread to worker thread
Summary: So-far-called `DatabaseThread` is not really database specific. It just handles a new thread + scheduling tasks. As with the upcoming e2e changes we want to use such threads also for the other things than databases, we should rename it so the name won't be confusing for anyone.

Test Plan: compile

Reviewers: palys-swm, ashoat

Reviewed By: ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1410