[server] Make sure we use string threadIDs in soft launch migration script
Summary: @atul noticed that some of the `source` fields in permission blobs contained `number`s instead of `string`s. This shouldn't affect anything, but it's good to be consistent.

Test Plan: Re-run the soft launch migration and make sure the `permissions` column in the `memberships` table has `number`s instead of `string`s

Reviewers: atul, palys-swm

Reviewed By: atul, palys-swm

Subscribers: KatPo, atul, Adrian

Differential Revision: https://phabricator.ashoat.com/D1411