[lib] Hide membership list of Genesis
Summary:
As described two diffs back, we want to disable relationship row creation in Genesis. Since members of Genesis will no longer have a guaranteed relationship with each other, it will be a bit misleading to enumerate all of them in the UI, since we will in fact only be showing the `UserInfo`s that the viewer has shared threads with.

Instead, we'll just avoid listing any members other than the current user and `ashoat` (the admin).

Test Plan: Make sure that I don't see the membership list in the iOS simulator after running `soft-launch-migration.js` on a clean database install

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1418