[native] Refactor error messages for worker threads
Summary:
Doing D1410 I missed the error messages that still indicate the operations are done on the database thread, while the rest of the code became more generic. To keep it consistent I think we should change those messages accordingly.
I was also thinking, maybe we should allow the developer to name worker threads? So we could then require a string name in the constructor of the worker thread and those error messages would contain that name so the developer would be informed clearly on which thread there was an error.
What do you think?

Test Plan: Honestly, none.

Reviewers: palys-swm, ashoat

Reviewed By: palys-swm, ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1421