[lib] Replace useCurrentThreadInfo with useExistingThreadInfoFinder
Summary:
Right now, the functionality to find an existing thread from a "search" (eg. a list of users being typed into the pending thread window) is contained within `useCurrentThreadInfo`, and requires updating `userInfoInputArray` to update.

I'm working on [this Notion issue](https://www.notion.so/commapp/Using-thread-composer-to-chat-with-non-friend-57748fac2a1b42b899e871b9e1b00e53), and I need a way to find an existing thread from a list of users within a callback. The reason is that I want that callback to set `setParams({ searching: false })`, but I can't do that until `baseThreadInfo` is updated, and that currently doesn't happen until `useCurrentThreadInfo` runs with the newest `userInfoInputArray`.

To make it possible to update the `baseThreadInfo` at the same time as we `setParams({ searching: false })`, I'm extracting this callback.

Test Plan: Tested `MessageListContainer` with pending threads: adding users, removing users, etc.

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1436