[native] Assign names to worker threads
Summary: In order to distinguish worker threads and make debugging easier we want to give them string name so we can tell on which thread an error occured exactly

Test Plan: We'd have to declare more worker threads, for now, we have just one - `database` used for SQLite. But you can always trigger an error on that database thread and see whether a proper error message is being displayed(with the thread name specified).

Reviewers: palys-swm, ashoat

Reviewed By: ashoat

Subscribers: ashoat, KatPo, palys-swm, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1423