[lib] Restrict solo "See sidebars..." in ThreadList
Summary:
The behavior today: before you join a sidebar in a thread, the thread appears without children in the thread list. After you join, the thread will always have a child when it appears in the thread list, but it might just be "See sidebars..." if all the sidebars (that the user is a member of)  have been read and not received any interaction in three days.

After this day: if no there is no sidebar (that the user is a member of) that is either unread or recent (activity in the past three days), then we will not show any children (no "See sidebars..."). If there is at least one, we will show "See more..." if there are any sidebars (that the user is a member of) that aren't been shown (the old behavior).

Test Plan: Test simulator and confirm fewer "See sidebars..."

Reviewers: palys-swm, atul

Reviewed By: palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1487