[native] Add an option to initialize measurer with precomputed heights
Summary: Determine initial state based on `measuredHeights` from the props. We need to compute `measurableItems` for all the items that are included in `initialMeasuredHeights`. This might result in all the items being computed, which would require triggering a callback, but this doesn't require any changes because `componentDidMount` already handles that.

Test Plan: Create a `NodeHeightMeasurer` with `initialMeasuredHeights`, check with console log if `measurableItems` were computed correctly and if `onMessagesMeasured` callback was called with correct values.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1506