[fix] Fix `log-in-panel` `validEmailRegex` check
Summary: The return value when calling `.search(validEmailRegex)` with non-email input is -1. However, -1 when cast to a boolean isn't falsey and leads to unintentionally blocking users from logging in with valid usernames.

Test Plan: Couldn't log in with valid username before change, can log in with valid username after change.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1497