[native] Extend allHeightsMeasured interface
Summary: We need a way to share `measuredHeights` between measurers so that we can initialize one with the result of another. To do that we need to extend a callback interface with additional parameter.

Test Plan: Modify a callback to console log the result and check if correct values are returned.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1504