[native] Focus ChatInputBar when non-friend selected
Summary:
When the user selects a non-friend as the first selection in the pending thread state (when `searching` is set), we will resolve to a pending thread with that user.

This diff makes it so when that happens, we also focus the `TextInput` in the `ChatInputBar`. We do that via `addReply('')`, which triggers `focusAndUpdateText`, but only if `ChatInputBar` is currently rendered. To make this work I had to make sure `ChatInputBar` is always rendered, so I made it invisible/unpressable in the cases where it was previously not being rendered.

On one hand this approach is a little sneaky, but it works perfectly and I actually think the mechanism is ideal.

Test Plan: Open up pending thread `searching` state, select non-friend, make sure `ChatInputBar` is focused

Reviewers: atul, palys-swm

Reviewed By: atul, palys-swm

Subscribers: KatPo, Adrian

Differential Revision: https://phabricator.ashoat.com/D1530