[lib] Merge ThreadMessageInfos when collision occurs
Summary: Refer to the [Notion issue](https://www.notion.so/commapp/Message-store-should-not-contain-a-new-thread-4da265b499d445c3840416322894376b). I regret not pushing harder on this in D926... the `invariant` here was very brittle and it was inevitable that it would eventually get tripped.

Test Plan: Repro the issue as described in the Notion task (without the preceding diff) and make sure it doesn't occur

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1533