[lib] Call reassignMessagesToRealizedThreads on processServerRequestsActionType
Summary: Refer to the [Notion issue](https://www.notion.so/commapp/Message-store-should-not-contain-a-new-thread-4da265b499d445c3840416322894376b). The state check mechanism can deliver realized `ThreadInfo`s into the `ThreadStore` via `processServerRequestsActionType`, so we should make sure to call `reassignMessagesToRealizedThreads` when that happens.

Test Plan: Repro the issue as described in the Notion task and make sure it doesn't occur

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1532