[lib] Increase login timeout to 30s
Summary: This is so I can log in on prod. It's only a temporary patch... hopefully we can soon prioritize (1) improving query perf, and (2) reducing the set of stuff that needs to be downloaded on a full state sync.

Test Plan: Make sure I can log in on prod with this patch

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1569