[native] Do not show a tooltip when there are no visible entries for a multimedia message
Summary: In the new tooltip highlight approach, we're going to have an option for a multimedia message to have no tooltip entries. In that case we don't want to navigate to an empty tooltip.

Test Plan: Modify visible entries function to return an empty array and verify that after long press a tooltip is not displayed.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1572