[native] Remove save action from multimedia tooltip
Summary: We decided to remove this action from a tooltip, because it is problematic when there are multiple images in one message and is available in ImageModal.

Test Plan: Long press on an image and check if only create / open sidebar action is displayed. Open a sidebar, long press the image - a tooltip should not appear.

Reviewers: ashoat

Reviewed By: ashoat

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1573