[native] Stop resetting navigation state on navigation logIn action
Summary:
The `logIn` action always occurs when the app is started. By resetting the navigation state in `logIn` we're clearing the calendar tab.

I think it should be enough to reset the navigation state on log out. We don't need to do it on log in.

Test Plan: Make sure problem is solved

Reviewers: atul

Reviewed By: atul

Subscribers: KatPo, palys-swm, Adrian

Differential Revision: https://phabricator.ashoat.com/D1597