[native] Fix bug where gestureEnabled accidentally gets set to true for Android
Summary: The logic in `MessageList` and `ThreadSettings` is meant to temporarily block gestures while a scroll-blocking modal is open, but when I introduced it I forgot to consider that we disable `gestureEnabled` on Android. As a result the logic was inadvertedly enabling gestures on Android when a scroll-blocking modal was open and then dismissed.

Test Plan: I made sure that the chat navigator would not have gestures enabled on Android after an overlay was displayed and then dismissed

Reviewers: palys-swm

Reviewed By: palys-swm

Subscribers: KatPo, Adrian, atul

Differential Revision: https://phabricator.ashoat.com/D1622